Module tor_cell::chancell

Messages sent over Tor channels

A ‘channel’ is a direct connection between a tor client and a relay, or between two relays. Current channels all use TLS.

This module implements the ChanCell type, which is the encoding for data sent over a channel. It also encodes and decodes various channel messages, which are the types of data conveyed over a channel.

Modules

codec

Implementation for encoding and decoding of ChanCells.

msg

Different kinds of messages that can be encoded in channel cells.

Structs

ChanCell

A decoded channel cell, to be sent or received on a channel.

ChanCmd

A ChanCmd is the type of a channel cell. The value of the ChanCmd indicates the meaning of the cell, and (possibly) its length.

CircId

Channel-local identifier for a circuit.

Constants

CELL_DATA_LEN

The amount of data sent in a fixed-length cell.

Traits

ChanMsg

Trait implemented by anything that can serve as a channel message.

Type Definitions

AnyChanCell

A decoded and parsed channel cell of unrestricted type.

BoxedCellBody

A RawCellBody stored on the heap.

RawCellBody

A cell body considered as a raw array of bytes