Module tor_cell::chancell::msg

Different kinds of messages that can be encoded in channel cells.

Structs

AuthChallenge

An AuthChallenge message is part of negotiation, sent by responders to initiators.

Authenticate

Part of negotiation: sent by initiators to responders.

Authorize

The Authorize message type is not yet used.

Certs

A Certs message is used as part of the channel handshake to send additional certificates.

Create

A Create message creates a circuit, using the TAP handshake.

Create2

A Create2 message create a circuit on the current channel.

CreateFast

A CreateFast message creates a circuit using no public-key crypto.

Created

A Created message responds to a Created message, using the TAP handshake.

Created2

A Created2 message completes a circuit-creation handshake.

CreatedFast

A CreatedFast message responds to a CreateFast message

Destroy

The Destroy message tears down a circuit.

DestroyReason

Declared reason for ending a circuit.

Netinfo

The netinfo message ends channel negotiation.

Padding

A Padding message is a fixed-length message on a channel that is ignored.

PaddingNegotiate

A PaddingNegotiate message is used to negotiate channel padding.

PaddingNegotiateCmd

A ChanCmd is the type of a channel cell. The value of the ChanCmd indicates the meaning of the cell, and (possibly) its length.

Relay

A Relay cell - that is, one transmitted over a circuit.

RelayEarly

A Relay cell that is allowed to contain a CREATE message.

Unrecognized

Holds any message whose command we don’t recognize.

Versions

A Versions message begins channel negotiation.

Vpadding

A VPadding message is a variable-length padding message.

Enums

AnyChanMsg

Decoded message from a channel.

Traits

Body

Trait for the ‘bodies’ of channel messages.