Enum tor_cell::chancell::msg::AnyChanMsg

#[non_exhaustive]
pub enum AnyChanMsg {

    Padding(Padding),
    Vpadding(Vpadding),
    Create(Create),
    CreateFast(CreateFast),
    Create2(Create2),
    Created(Created),
    CreatedFast(CreatedFast),
    Created2(Created2),
    Relay(Relay),
    RelayEarly(RelayEarly),
    Destroy(Destroy),
    Netinfo(Netinfo),
    Versions(Versions),
    PaddingNegotiate(PaddingNegotiate),
    Certs(Certs),
    AuthChallenge(AuthChallenge),
    Authenticate(Authenticate),
    Authorize(Authorize),
    Unrecognized(Unrecognized),

}

Decoded message from a channel.

A ChanMsg is an item received on a channel – a message from another Tor client or relay that we are connected to directly over a TLS connection.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Padding(Padding)

A Padding message

Vpadding(Vpadding)

Variable-length padding message

Create(Create)

(Deprecated) TAP-based cell to create a new circuit.

CreateFast(CreateFast)

(Mostly deprecated) HMAC-based cell to create a new circuit.

Create2(Create2)

Cell to create a new circuit

Created(Created)

(Deprecated) Answer to a Create cell

CreatedFast(CreatedFast)

(Mostly Deprecated) Answer to a CreateFast cell

Created2(Created2)

Answer to a Create2 cell

Relay(Relay)

A message sent along a circuit, likely to a more-distant relay.

RelayEarly(RelayEarly)

A message sent along a circuit (limited supply)

Destroy(Destroy)

Tear down a circuit

Netinfo(Netinfo)

Part of channel negotiation: describes our position on the network

Versions(Versions)

Part of channel negotiation: describes what link protocol versions we support

PaddingNegotiate(PaddingNegotiate)

Negotiates what kind of channel padding to send

Certs(Certs)

Part of channel negotiation: additional certificates not in the TLS handshake

AuthChallenge(AuthChallenge)

Part of channel negotiation: additional random material to be used as part of authentication

Authenticate(Authenticate)

Part of channel negotiation: used to authenticate relays when they initiate the channel.

Authorize(Authorize)

Not yet used

Unrecognized(Unrecognized)

Any cell whose command we don’t recognize