Trait tor_cell::chancell::msg::Body

pub trait Body: Readable {
    // Required method
    fn encode_onto<W: Writer + ?Sized>(self, w: &mut W) -> EncodeResult<()>;

    // Provided method
    fn decode_from_reader(r: &mut Reader<'_>) -> Result<Self> { ... }
}

Trait for the ‘bodies’ of channel messages.

Required Methods

fn encode_onto<W: Writer + ?Sized>(self, w: &mut W) -> EncodeResult<()>

Consume this message and encode its body onto w.

Does not encode anything but the cell body, and does not pad to the cell length.

Provided Methods

fn decode_from_reader(r: &mut Reader<'_>) -> Result<Self>

Decode a channel cell body from a provided reader.