Module tor_cell::relaycell::hs

Encoding and decoding for relay messages related to onion services.

Modules

est_intro

Define the ESTABLISH_INTRO message and related types.

intro_payload

The encrypted portion of an INTRODUCE{1,2} message.

Structs

AuthKeyType

The type of the introduction point auth key

EstablishRendezvous

A message sent from client to rendezvous point.

IntroEstablished

Reply sent from the introduction point to the onion service, telling it that an introduction point is now established.

IntroEstablishedExtType

The recognized extension types for an IntroEstablished message.

Introduce1

A message sent from client to introduction point.

Introduce2

A message sent from introduction point to hidden service host.

IntroduceAck

A reply from the introduction point to the client, telling it that its introduce1 was received.

IntroduceAckExtType

The recognized extension types for an IntroEstablished message.

IntroduceAckStatus

A status code returned in response to an INTRODUCE1 message.

IntroduceExtType

The recognized extension types for an Introduce1 or `Introduce2 message.

IntroduceHeader

The unencrypted header portion of an Introduce1 or Introduce2 message.

Rendezvous1

A message sent from an onion service to a rendezvous point, telling it to make a connection to the client.

Rendezvous2

A message sent from the rendezvous point to the client, telling it about the onion service’s message.

RendezvousEstablished

Acknowledges an EstablishRendezvous message.

UnrecognizedExt

An unrecognized extension for some HS-related message.