Module tor_cell::relaycell::msg

Encoding and decoding for relay messages

Relay messages are sent along circuits, inside RELAY or RELAY_EARLY cells.

Re-exports

pub use super::hs::est_intro::EstablishIntro;

pub use super::hs::EstablishRendezvous;

pub use super::hs::IntroEstablished;

pub use super::hs::Introduce1;

pub use super::hs::Introduce2;

pub use super::hs::IntroduceAck;

pub use super::hs::Rendezvous1;

pub use super::hs::Rendezvous2;

pub use super::hs::RendezvousEstablished;

Structs

Begin

A Begin message creates a new data stream.

BeginDir

Opens a new stream on a directory cache.

BeginFlags

A set of recognized flags that can be attached to a begin cell.

Connected

A Connected message is a successful response to a Begin message

Data

A Data message represents data sent along a stream.

Drop

A padding message, which is always ignored.

End

An End message tells the other end of the circuit to close a stream.

EndReason

A declared reason for closing a stream

Extend

Extend was an obsolete circuit extension message format.

Extend2

An Extend2 message tells the last relay in a circuit to extend to a new hop.

Extended

Extended was an obsolete circuit extension message, sent in reply to an Extend message.

Extended2

Extended2 is a successful reply to an Extend2 message.

Resolve

A Resolve message launches a DNS lookup stream.

Resolved

A Resolved message is a successful reply to a Resolve message.

Sendme

A Sendme message is used to increase flow-control windows.

Truncate

Tells a circuit to close all downstream hops on the circuit.

Truncated

A Truncated message is sent to the client when the remaining hops of a circuit have gone away.

Unrecognized

A relay message that we didn’t recognize

Enums

AnyRelayMsg

A single parsed relay message, sent or received along a circuit

IpVersionPreference

A preference for IPv4 vs IPv6 addresses; usable as a nicer frontend for BeginFlags.

ResolvedVal

Possible response to a DNS lookup

Traits

Body

Internal: traits in common different cell bodies.