Enum tor_cell::relaycell::msg::AnyRelayMsg

#[non_exhaustive]
pub enum AnyRelayMsg {

    Begin(Begin),
    Data(Data),
    End(End),
    Connected(Connected),
    Sendme(Sendme),
    Extend(Extend),
    Extended(Extended),
    Extend2(Extend2),
    Extended2(Extended2),
    Truncate(Truncate),
    Truncated(Truncated),
    Drop(Drop),
    Resolve(Resolve),
    Resolved(Resolved),
    BeginDir(BeginDir),
    EstablishIntro(EstablishIntro),
    EstablishRendezvous(EstablishRendezvous),
    Introduce1(Introduce1),
    Introduce2(Introduce2),
    Rendezvous1(Rendezvous1),
    Rendezvous2(Rendezvous2),
    IntroEstablished(IntroEstablished),
    RendezvousEstablished(RendezvousEstablished),
    IntroduceAck(IntroduceAck),
    Unrecognized(Unrecognized),

}

A single parsed relay message, sent or received along a circuit

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

Begin(Begin)

Create a stream

Data(Data)

Send data on a stream

End(End)

Close a stream

Connected(Connected)

Successful response to a Begin message

Sendme(Sendme)

For flow control

Extend(Extend)

Extend a circuit to a new hop (deprecated)

Extended(Extended)

Successful response to an Extend message (deprecated)

Extend2(Extend2)

Extend a circuit to a new hop

Extended2(Extended2)

Successful response to an Extend2 message

Truncate(Truncate)

Partially close a circuit

Truncated(Truncated)

Tell the client that a circuit has been partially closed

Drop(Drop)

Used for padding

Resolve(Resolve)

Launch a DNS request

Resolved(Resolved)

Response to a Resolve message

BeginDir(BeginDir)

Start a directory stream

EstablishIntro(EstablishIntro)

Establish Introduction

EstablishRendezvous(EstablishRendezvous)

Establish Rendezvous

Introduce1(Introduce1)

Introduce1 (client to introduction point)

Introduce2(Introduce2)

Introduce2 (introduction point to service)

Rendezvous1(Rendezvous1)

Rendezvous1 (service to rendezvous point)

Rendezvous2(Rendezvous2)

Rendezvous2 (rendezvous point to client)

IntroEstablished(IntroEstablished)

Acknowledgement for EstablishIntro.

RendezvousEstablished(RendezvousEstablished)

Acknowledgment for EstablishRendezvous.

IntroduceAck(IntroduceAck)

Acknowledgement for Introduce1.

Unrecognized(Unrecognized)

An unrecognized command.