Trait tor_cell::relaycell::msg::Body

pub trait Body: Sized {
    // Required methods
    fn decode_from_reader(r: &mut Reader<'_>) -> Result<Self>;
    fn encode_onto<W: Writer + ?Sized>(self, w: &mut W) -> EncodeResult<()>;
}

Internal: traits in common different cell bodies.

Required Methods

fn decode_from_reader(r: &mut Reader<'_>) -> Result<Self>

Decode a relay cell body from a provided reader.

fn encode_onto<W: Writer + ?Sized>(self, w: &mut W) -> EncodeResult<()>

Encode the body of this cell into the end of a writer.